pip-venv-requirements.txt
jmespath==0.10.0
ansible-core==2.14.18
jinja2==3.1.2
ansible-lint
httpx
PyYAML
